% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{log_like_Bernoulli}
\alias{log_like_Bernoulli}
\title{Bernoulli Log Likelihood}
\usage{
log_like_Bernoulli(x, theta, q)
}
\arguments{
\item{x}{matrix with all binary entries}

\item{theta}{estimated natural parameters with 
same dimensions as x}

\item{q}{instead of x, you can input matrix q which is 
-1 if \code{x = 0}, 1 if \code{x = 1}, and 0 if \code{is.na(x)}}
}
\description{
Calculate the Bernoulli log likelihood of matrix
}

