\name{logisticPCA-package}
\alias{logisticPCA-package}
\alias{logisticPCA}
\docType{package}
\title{
logitsticPCA
}
\description{
Calculates logistic PCA on binary data
}
\details{
\tabular{ll}{
Package: \tab logisticPCA\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-12-08\cr
License: \tab MIT\cr
}

}
\author{
Andrew Landgraf

Maintainer: Andrew Landgraf <landgraf.5@osu.edu>
}
\references{
TBA
}
\keyword{ package }
\seealso{
}
\examples{
}
