% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convexLogisticPCA.R
\name{plot.clpca}
\alias{plot.clpca}
\title{Plot convex logistic PCA}
\usage{
\method{plot}{clpca}(x, type = c("trace", "loadings", "scores"), ...)
}
\arguments{
\item{x}{convex logistic PCA object}

\item{type}{the type of plot \code{type = "trace"} plots the algorithms progress by
iteration, \code{type = "loadings"} plots the first 2 PC loadings,
\code{type = "scores"} plots the first 2 PC scores}

\item{...}{Additional arguments}
}
\description{
Plots the results of a convex logistic PCA
}
\examples{
# construct a low rank matrix in the logit scale
rows = 100
cols = 10
set.seed(1)
mat_logit = outer(rnorm(rows), rnorm(cols))

# generate a binary matrix
mat = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit)) * 1.0

# run convex logistic PCA on it
clpca = convexLogisticPCA(mat, k = 2, M = 4, main_effects = FALSE)

\dontrun{
plot(clpca)
}
}

