% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logisticPCA.R
\name{print.lpca}
\alias{print.lpca}
\title{Print logistic PCA object}
\usage{
\method{print}{lpca}(x, ...)
}
\arguments{
\item{x}{logistic PCA object}

\item{...}{Additional arguments}
}
\description{
Print logistic PCA object
}

