% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logisticSVD.R
\name{print.lsvd}
\alias{print.lsvd}
\title{Print logistic SVD object}
\usage{
\method{print}{lsvd}(x, ...)
}
\arguments{
\item{x}{logistic SVD object}

\item{...}{Additional arguments}
}
\description{
Print logistic SVD object
}

