% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logisticPCA-package.R
\docType{package}
\name{logisticPCA-package}
\alias{logisticPCA-package}
\title{logisticPCA-package}
\description{
Dimension reduction techniques for binary data including logistic PCA
}
\author{
Andrew Landgraf
}
\keyword{package}

