% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.lsvd}
\alias{plot.lsvd}
\title{Plot logistic SVD}
\usage{
\method{plot}{lsvd}(object, type = c("trace", "loadings"), ...)
}
\arguments{
\item{object}{logistic SVD object}

\item{type}{the type of plot \code{type = "trace"} plots the algorithms progress by
iteration, \code{type = "loadings"} plots the loadings first 2 principal components}

\item{...}{Additional arguments}
}
\description{
Plots the results of a logistic SVD
}
\examples{
# construct a low rank matrix in the logit scale
rows = 100
cols = 10
set.seed(1)
mat_logit = outer(rnorm(rows), rnorm(cols))

# generate a binary matrix
mat = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit)) * 1.0

# run logistic SVD on it
lsvd = logisticSVD(mat, k = 2, main_effects = FALSE, use_irlba = FALSE)

\dontrun{
plot(lsvd)
}
}

