% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.lpca}
\alias{predict.lpca}
\title{Predict Logistic PCA scores on new data}
\usage{
\method{predict}{lpca}(object, newdata)
}
\arguments{
\item{object}{A logistic PCA object}

\item{newdata}{Binary matrix to apply logistic PCA on. If missing, will return PCs
that the data \code{object} was fit on}
}
\description{
Predict Logistic PCA scores on new data
}
\examples{
# construct a low rank matrices in the logit scale
rows = 100
cols = 10
loadings = rnorm(cols)
mat_logit = outer(rnorm(rows), loadings)
mat_logit_new = outer(rnorm(rows), loadings)

# convert to a binary matrix
mat = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit)) * 1.0
mat_new = (matrix(runif(rows * cols), rows, cols) <= inv.logit.mat(mat_logit_new)) * 1.0

# run logistic PCA on it
lpca = logisticPCA(mat, k = 1, M = 4, main_effects = FALSE)

PCs = predict(lpca, mat_new)
}

