% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.clpca}
\alias{print.clpca}
\title{Print convex logistic PCA object}
\usage{
\method{print}{clpca}(x, ...)
}
\arguments{
\item{x}{convex logistic PCA object}

\item{...}{Additional arguments}
}
\description{
Print convex logistic PCA object
}

