% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexLogisticPCA.R
\name{project.Fantope}
\alias{project.Fantope}
\title{Project onto the Fantope}
\usage{
project.Fantope(x, k)
}
\arguments{
\item{x}{a symmetric matrix}

\item{k}{the rank of the Fantope desired}
}
\value{
\item{H}{a rank \code{k} Fantope matrix}
\item{U}{a \code{k}-dimentional orthonormal matrix with the first \code{k} eigenvectors of \code{H}}
}
\description{
Project a symmetric matrix onto the convex set of the rank k Fantope
}

